2011-11-09 Roman - UI.Vision: Falgs store for developed widgets is changed to uint8_t type.     Child attributes process into change rule "attr" is added.     Rule's name process is added for group attributes change into one rule "attr".
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1637 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/ui/Vision/vis_devel_widgs.h
@@ -558,18 +558,18 @@ class DevelWdgView: public WdgView
 	void upMouseCursors( const QPoint &pnt );
 
 	//Private attributes
-	short int	fMakeScale	:1;	//Make visual item scaling
-	short int	fWdgEdit	:1;	//Widget under edition
-	short int	fWdgSelect	:1;	//Widget is selected
-	short int	fMoveHold	:1;	//Mouse move hold state
-	short int	fHoldChild	:1;	//Hold child widget in time of moving and resizing
-	short int	fLeftTop	:1;	//Left top anchors
-	short int	fHoldSelRect	:1;	//Hold for select rect
-	short int	fMoveHoldMove	:1;	//Mouse move on hold
-	short int	fHideChilds	:1;	//Hide childs on move
-	short int	fSelChange	:1;	//Changed select map
-        short int	fPrevEdExitFoc  :1;	//Prevention exit from widget edition by focus loosing
-        short int	fFocus		:1;	//Edition window of the widget in focus
+	uint8_t	fMakeScale	:1;	//Make visual item scaling
+        uint8_t	fWdgEdit	:1;	//Widget under edition
+        uint8_t	fWdgSelect	:1;	//Widget is selected
+        uint8_t	fMoveHold	:1;	//Mouse move hold state
+        uint8_t	fHoldChild	:1;	//Hold child widget in time of moving and resizing
+        uint8_t	fLeftTop	:1;	//Left top anchors
+        uint8_t	fHoldSelRect	:1;	//Hold for select rect
+        uint8_t	fMoveHoldMove	:1;	//Mouse move on hold
+        uint8_t	fHideChilds	:1;	//Hide childs on move
+        uint8_t	fSelChange	:1;	//Changed select map
+        uint8_t	fPrevEdExitFoc  :1;	//Prevention exit from widget edition by focus loosing
+        uint8_t	fFocus		:1;	//Edition window of the widget in focus
 
 	float		mVisScale;		//Visual scale value of root widget.
